f + h instead of f?
I am pretty sure in the current version  at least iff does not stand. Any f not identically 0 would have a larger hilbert norm than any h that is identically 0.

On the other hand, if we were to show ||f + h||_{H} \geq ||f||_{h}, this indeed holds via direct expansion, and is what are looking for in the later part.